refactor: move row_matrix in a separate file
